feat: optional GUI renderer
